requirements.txt
Django==5.0.4
python-dotenv==1.0.1
psycopg2==2.9.9
pillow==10.3.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
